refactor: implement explosion variant prop
