+++ packages/client/src/game/entities/Explosion.ts
@@ -1,24 +1,26 @@
 import { spriteCoordinates } from '../data/constants';
-import { EntityEvent } from './../typings/index';
+import { EntityEvent, ExplosionVariant } from '../typings';
 import { Entity, Projectile, Tank } from './';
 
 type ExplosionSettings = { parentEntity: Tank | Projectile };
 
 export class Explosion extends Entity {
+  variant: ExplosionVariant = 'PROJECTILE_EXPLOSION';
+
   constructor(props: ExplosionSettings) {
     super({ posX: 0, posY: 0 });
     Object.assign(this, this.calculateExplosionProps(props.parentEntity));
-    this.role = 'neutral';
+    this.type = 'explosion';
     this.crossable = true;
     this.hittable = false;
     this.color = 'transparent';
 
     let delay = 0;
-    switch (this.type) {
-      case 'projectileExplosion':
+    switch (this.variant) {
+      case 'PROJECTILE_EXPLOSION':
         this.mainSpriteCoordinates = spriteCoordinates.projectileExplosion;
         break;
-      case 'tankExplosion':
+      case 'TANK_EXPLOSION':
         this.mainSpriteCoordinates = spriteCoordinates.tankExplosion;
         delay = 16; // для более красивой отрисовки взрыва танка нужна задержка
         break;
@@ -39,11 +41,11 @@ export class Explosion extends Entity {
 
   calculateExplosionProps(entity: Tank | Projectile) {
     const size = entity.type === 'projectile' ? 4 : 8;
-    const type = entity.type === 'projectile' ? 'projectileExplosion' : 'tankExplosion';
+    const variant = entity.type === 'projectile' ? 'PROJECTILE_EXPLOSION' : 'TANK_EXPLOSION';
     let posX = entity.posX;
     let posY = entity.posY;
 
-    /**Без коррекции взрыв рисуется не по центру обьекта в который попал снаряд 
+    /** Без коррекции взрыв рисуется не по центру обьекта в который попал снаряд 
     из-за несовпадения координат и разницы в размерах взрыва и обьекта. */
     const correction = entity.direction === 'UP' || entity.direction === 'LEFT' ? -2 : 0;
 
@@ -77,6 +79,6 @@ export class Explosion extends Entity {
       }
     }
 
-    return { type, posX, posY, width: size, height: size };
+    return { variant, posX, posY, width: size, height: size };
   }
 }
